temporary drop metrics tests while speeding them up
